requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
argon2-cffi==25.1.0
httpx==0.28.1
